no-mistakes(review): centralize test module stubs in tests/conftest.py
